Flatten api directory
