+++ api/log-entries-id.ts
@@ -1,4 +1,4 @@
-import getAirtableContext from "../lib/airtableBase";
+import getAirtableContext from "./airtableBase";
 
 const idLogEntryHandler = async (req: any, res: any) => {
   const { base, TABLES, airtableToken, baseId } = getAirtableContext();
