+++ api/snapshots-id.ts
@@ -1,29 +1,18 @@
-import getAirtableContext from "../lib/airtableBase";
+import getAirtableContext from "./airtableBase";
 
-const idThreadsHandler = async (req: any, res: any) => {
+const idSnapshotsHandler = async (req: any, res: any) => {
     const { base, TABLES, airtableToken, baseId } = getAirtableContext();
 
     const { id } = req.query;
     if (!id) {
-        return res.status(400).json({ error: "Missing thread ID" });
+        return res.status(400).json({ error: "Missing snapshot ID" });
     }
 
-    if (!airtableToken) {
-        console.error("Missing AIRTABLE_API_KEY");
-        return res.status(500).json({ error: "Missing Airtable API Key" });
+    if (!airtableToken || !baseId || !TABLES.SNAPSHOTS) {
+        return res.status(500).json({ error: "Missing Airtable configuration" });
     }
 
-    if (!baseId) {
-        console.error("Missing AIRTABLE_BASE_ID");
-        return res.status(500).json({ error: "Missing Airtable Base ID" });
-    }
-
-    if (!TABLES?.THREADS) {
-        console.error("Missing AIRTABLE_THREADS_TABLE_NAME");
-        return res.status(500).json({ error: "Missing Airtable Threads Table Name" });
-    }
-
-    const recordUrl = `https://api.airtable.com/v0/${baseId}/${encodeURIComponent(TABLES.THREADS)}/${id}`;
+    const recordUrl = `https://api.airtable.com/v0/${baseId}/${encodeURIComponent(TABLES.SNAPSHOTS)}/${id}`;
 
     const config = {
         headers: {
@@ -70,4 +59,4 @@ const idThreadsHandler = async (req: any, res: any) => {
     }
 };
 
-export default idThreadsHandler;
+export default idSnapshotsHandler;
