+++ api/synthesisUtils.ts
